DRY coordinate parser for rotation position and scale
