+++ src/components/position.js
@@ -1,6 +1,8 @@
+var coordinateParser = require('./coordinate-parser');
 var registerComponent = require('../core/register-component').registerComponent;
+var utils = require('../vr-utils');
 
-module.exports.Component = registerComponent('position', {
+var proto = {
   defaults: {
     value: {
       x: 0,
@@ -16,26 +18,8 @@ module.exports.Component = registerComponent('position', {
       // Updates three.js object
       object3D.position.set(data.x, data.y, data.z);
     }
-  },
-
-  parseAttributesString: {
-    value: function (attrs) {
-      var defaults = this.defaults;
-      if (typeof attrs !== 'string') { return attrs; }
-      var values = attrs.split(' ');
-      return {
-        x: parseFloat(values[0] || defaults.x),
-        y: parseFloat(values[1] || defaults.y),
-        z: parseFloat(values[2] || defaults.z)
-      };
-    }
-  },
-
-  stringifyAttributes: {
-    value: function (attrs) {
-      if (typeof attrs !== 'object') { return attrs; }
-      return [attrs.x, attrs.y, attrs.z].join(' ');
-    }
   }
+};
 
-});
+utils.mixin(proto, coordinateParser);
+module.exports.Component = registerComponent('position', proto);
