+++ src/components/rotation.js
@@ -1,7 +1,9 @@
+var coordinateParser = require('./coordinate-parser');
 var registerComponent = require('../core/register-component').registerComponent;
 var THREE = require('../../lib/three');
+var utils = require('../vr-utils');
 
-module.exports.Component = registerComponent('rotation', {
+var proto = {
   defaults: {
     value: {
       x: 0,
@@ -21,25 +23,8 @@ module.exports.Component = registerComponent('rotation', {
       // Updates three.js object
       object3D.rotation.set(rotationX, rotationY, rotationZ);
     }
-  },
-
-  parseAttributesString: {
-    value: function (attrs) {
-      var defaults = this.defaults;
-      if (typeof attrs !== 'string') { return attrs; }
-      var values = attrs.split(' ');
-      return {
-        x: parseFloat(values[0] || defaults.x),
-        y: parseFloat(values[1] || defaults.y),
-        z: parseFloat(values[2] || defaults.z)
-      };
-    }
-  },
-
-  stringifyAttributes: {
-    value: function (attrs) {
-      if (typeof attrs !== 'object') { return attrs; }
-      return [attrs.x, attrs.y, attrs.z].join(' ');
-    }
   }
-});
+};
+
+utils.mixin(proto, coordinateParser);
+module.exports.Component = registerComponent('rotation', proto);
