+++ src/components/scale.js
@@ -1,6 +1,8 @@
+var coordinateParser = require('./coordinate-parser');
 var registerComponent = require('../core/register-component').registerComponent;
+var utils = require('../vr-utils');
 
-module.exports.Component = registerComponent('scale', {
+var proto = {
   defaults: {
     value: {
       x: 1,
@@ -15,25 +17,8 @@ module.exports.Component = registerComponent('scale', {
       var object3D = this.el.object3D;
       object3D.scale.set(data.x, data.y, data.z);
     }
-  },
-
-  parseAttributesString: {
-    value: function (attrs) {
-      var defaults = this.defaults;
-      if (typeof attrs !== 'string') { return attrs; }
-      var values = attrs.split(' ');
-      return {
-        x: parseFloat(values[0] || defaults.x),
-        y: parseFloat(values[1] || defaults.y),
-        z: parseFloat(values[2] || defaults.z)
-      };
-    }
-  },
-
-  stringifyAttributes: {
-    value: function (attrs) {
-      if (typeof attrs !== 'object') { return attrs; }
-      return [attrs.x, attrs.y, attrs.z].join(' ');
-    }
   }
-});
+};
+
+utils.mixin(proto, coordinateParser);
+module.exports.Component = registerComponent('scale', proto);
